Change in number of files by mean
The files by label was sampled by the mean of files contained in each data set.  When the label folder has less files, all files were selected, when has more files, a random sample was selected and no more files than mean.  The same was did with first and third quartile, an arbitrary quantity of 1,500 files.